refactor: extract PDF region block recovery
